Support for big PDFs over 2G. (Initial contribution by Welman Jordan)
git-svn-id: https://svn.code.sf.net/p/itext/code/trunk@5017 3a709846-db2a-4fbd-be0b-eb932eee6479

+++ itext/src/main/java/com/itextpdf/text/pdf/CFFFont.java
@@ -248,7 +248,7 @@ public class CFFFont {
 
     int getPosition() {
         try {
-            return buf.getFilePointer();
+            return (int)buf.getFilePointer();
         }
         catch (Exception e) {
             throw new ExceptionConverter(e);
