+++ itext/src/main/java/com/itextpdf/text/pdf/EnumerateTTC.java
@@ -78,7 +78,7 @@ class EnumerateTTC extends TrueTypeFont{
             rf.skipBytes(4);
             int dirCount = rf.readInt();
             names = new String[dirCount];
-            int dirPos = rf.getFilePointer();
+            int dirPos = (int)rf.getFilePointer();
             for (int dirIdx = 0; dirIdx < dirCount; ++dirIdx) {
                 tables.clear();
                 rf.seek(dirPos);
